+++ projects/portal/src/app/models/cosmos/distribution.service.ts
@@ -1,0 +1,112 @@
+import { CosmosSDKService } from '../cosmos-sdk.service';
+import { KeyService } from '../keys/key.service';
+import { CosmosWallet } from '../wallets/wallet.model';
+import { SimulatedTxResultResponse } from './tx-common.model';
+import { TxCommonService } from './tx-common.service';
+import { Injectable } from '@angular/core';
+import { cosmosclient, rest, proto } from '@cosmos-client/core';
+import { InlineResponse20075 } from '@cosmos-client/core/esm/openapi';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class DistributionService {
+  constructor(
+    private readonly cosmosSDK: CosmosSDKService,
+    private readonly key: KeyService,
+    private readonly txCommonService: TxCommonService,
+  ) {}
+
+  // Withdraw Delegator Reward
+  async withdrawDelegatorReward(
+    validatorAddress: string,
+    currentCosmosWallet: CosmosWallet,
+    gas: proto.cosmos.base.v1beta1.ICoin,
+    fee: proto.cosmos.base.v1beta1.ICoin,
+    privateKey?: string,
+  ): Promise<InlineResponse20075> {
+    const cosmosPublicKey = currentCosmosWallet.public_key;
+    const txBuilder = await this.buildWithdrawDelegatorRewardTxBuilder(
+      validatorAddress,
+      cosmosPublicKey,
+      gas,
+      fee,
+    );
+    const signerBaseAccount = await this.txCommonService.getBaseAccount(cosmosPublicKey);
+    if (!signerBaseAccount) {
+      throw Error('Unsupported Account!');
+    }
+    const signedTxBuilder = await this.txCommonService.signTx(
+      txBuilder,
+      signerBaseAccount,
+      currentCosmosWallet,
+      privateKey,
+    );
+    if (!signedTxBuilder) {
+      throw Error('Failed to sign!');
+    }
+    const txResult = await this.txCommonService.announceTx(signedTxBuilder);
+    return txResult;
+  }
+
+  async simulateToWithdrawDelegatorReward(
+    validatorAddress: string,
+    cosmosPublicKey: cosmosclient.PubKey,
+    minimumGasPrice: proto.cosmos.base.v1beta1.ICoin,
+    gasRatio: number,
+  ): Promise<SimulatedTxResultResponse> {
+    const dummyFee: proto.cosmos.base.v1beta1.ICoin = {
+      denom: minimumGasPrice.denom,
+      amount: '1',
+    };
+    const dummyGas: proto.cosmos.base.v1beta1.ICoin = {
+      denom: minimumGasPrice.denom,
+      amount: '1',
+    };
+    const simulatedTxBuilder = await this.buildWithdrawDelegatorRewardTxBuilder(
+      validatorAddress,
+      cosmosPublicKey,
+      dummyGas,
+      dummyFee,
+    );
+    return await this.txCommonService.simulateTx(simulatedTxBuilder, minimumGasPrice, gasRatio);
+  }
+
+  async buildWithdrawDelegatorRewardTxBuilder(
+    validatorAddress: string,
+    cosmosPublicKey: cosmosclient.PubKey,
+    gas: proto.cosmos.base.v1beta1.ICoin,
+    fee: proto.cosmos.base.v1beta1.ICoin,
+  ): Promise<cosmosclient.TxBuilder> {
+    const baseAccount = await this.txCommonService.getBaseAccount(cosmosPublicKey);
+    if (!baseAccount) {
+      throw Error('Unused Account or Unsupported Account Type!');
+    }
+    const fromAddress = cosmosclient.AccAddress.fromPublicKey(cosmosPublicKey);
+    const msgWithdrawDelegatorReward = this.buildMsgWithdrawDelegatorReward(
+      fromAddress.toString(),
+      validatorAddress,
+    );
+
+    const txBuilder = await this.txCommonService.buildTxBuilder(
+      [msgWithdrawDelegatorReward],
+      cosmosPublicKey,
+      baseAccount,
+      gas,
+      fee,
+    );
+    return txBuilder;
+  }
+
+  buildMsgWithdrawDelegatorReward(
+    delegatorAddress: string,
+    validatorAddress: string,
+  ): proto.cosmos.distribution.v1beta1.MsgWithdrawDelegatorReward {
+    const msgWithdrawDelegatorReward =
+      new proto.cosmos.distribution.v1beta1.MsgWithdrawDelegatorReward({
+        delegator_address: delegatorAddress,
+        validator_address: validatorAddress,
+      });
+    return msgWithdrawDelegatorReward;
+  }
+}
