+++ projects/portal/src/app/models/cosmos/staking.application.service.ts
@@ -2,7 +2,6 @@ import { DelegateFormDialogComponent } from '../../pages/dialogs/delegate/delega
 import { DelegateMenuDialogComponent } from '../../pages/dialogs/delegate/delegate-menu-dialog/delegate-menu-dialog.component';
 import { RedelegateFormDialogComponent } from '../../pages/dialogs/delegate/redelegate-form-dialog/redelegate-form-dialog.component';
 import { UndelegateFormDialogComponent } from '../../pages/dialogs/delegate/undelegate-form-dialog/undelegate-form-dialog.component';
-import { WithdrawDelegatorRewardFormDialogComponent } from '../../pages/dialogs/delegate/withdraw-delegator-reward-form-dialog/withdraw-delegator-reward-form-dialog.component';
 import { convertHexStringToUint8Array } from '../../utils/converter';
 import { validatePrivateStoredWallet } from '../../utils/validater';
 import { TxFeeConfirmDialogComponent } from '../../views/cosmos/tx-fee-confirm-dialog/tx-fee-confirm-dialog.component';
@@ -75,14 +74,6 @@ export class StakingApplicationService {
   async openUndelegateFormDialog(validator: InlineResponse20066Validators): Promise<void> {
     const txHash = await this.dialog
       .open(UndelegateFormDialogComponent, { data: validator })
-      .afterClosed()
-      .toPromise();
-    await this.router.navigate(['txs', txHash]);
-  }
-
-  async openWithdrawDelegatorRewardFormDialog(validator: InlineResponse20066Validators): Promise<void> {
-    const txHash = await this.dialog
-      .open(WithdrawDelegatorRewardFormDialogComponent, { data: validator })
       .afterClosed()
       .toPromise();
     await this.router.navigate(['txs', txHash]);
@@ -561,95 +552,5 @@ export class StakingApplicationService {
 
     return txHash;
     // await this.router.navigate(['txs', txHash]);
-  }
-
-  async withdrawDelegatorReward(
-    validatorAddress: string,
-    minimumGasPrice: proto.cosmos.base.v1beta1.ICoin,
-    gasRatio: number,
-  ) {
-    // get public key
-    const currentCosmosWallet = await this.walletService.currentCosmosWallet$
-      .pipe(take(1))
-      .toPromise();
-    if (!currentCosmosWallet) {
-      throw Error('Current connected wallet is invalid!');
-    }
-    const cosmosPublicKey = currentCosmosWallet.public_key;
-    if (!cosmosPublicKey) {
-      throw Error('Invalid public key!');
-    }
-
-    // simulate
-    let simulatedResultData: SimulatedTxResultResponse;
-    let gas: proto.cosmos.base.v1beta1.ICoin;
-    let fee: proto.cosmos.base.v1beta1.ICoin;
-
-    const dialogRefSimulating = this.loadingDialog.open('Simulating...');
-
-    try {
-      simulatedResultData = await this.staking.simulateToWithdrawDelegatorReward(
-        validatorAddress,
-        cosmosPublicKey,
-        minimumGasPrice,
-        gasRatio,
-      );
-      gas = simulatedResultData.estimatedGasUsedWithMargin;
-      fee = simulatedResultData.estimatedFeeWithMargin;
-    } catch (error) {
-      console.error(error);
-      const errorMessage = `Tx simulation failed: ${(error as Error).toString()}`;
-      this.snackBar.open(`An error has occur: ${errorMessage}`, 'Close');
-      return;
-    } finally {
-      dialogRefSimulating.close();
-    }
-
-    // confirm fee only ununifi wallet type case
-    if (currentCosmosWallet.type === WalletType.ununifi) {
-      const txFeeConfirmedResult = await this.dialog
-        .open(TxFeeConfirmDialogComponent, {
-          data: {
-            fee,
-            isConfirmed: false,
-          },
-        })
-        .afterClosed()
-        .toPromise();
-      if (txFeeConfirmedResult === undefined || txFeeConfirmedResult.isConfirmed === false) {
-        this.snackBar.open('Tx was canceled', undefined, { duration: 6000 });
-        return;
-      }
-    }
-
-    // send tx
-    const dialogRef = this.loadingDialog.open('Sending');
-
-    let txResult: InlineResponse20075 | undefined;
-    let txHash: string | undefined;
-
-    try {
-      txResult = await this.staking.withdrawDelegatorReward(
-        validatorAddress,
-        currentCosmosWallet,
-        gas,
-        fee,
-      );
-      txHash = txResult.tx_response?.txhash;
-      if (txHash === undefined) {
-        throw Error('Invalid txHash!');
-      }
-    } catch (error) {
-      console.error(error);
-      const msg = (error as Error).toString();
-      this.snackBar.open(`An error has occur: ${msg}`, 'Close');
-      return;
-    } finally {
-      dialogRef.close();
-    }
-
-    this.snackBar.open('Successfully delegate to the validator', undefined, { duration: 6000 });
-
-    return txHash;
   }
 }
